+++ runners/utility.ts
@@ -59,6 +59,53 @@ export async function* readerToLines(
 }
 
 /**
+ * Convert an `AsyncIterable<Uint8Array>` into an `AsyncIterable<string>` of lines.
+ * @param buffs The iterable bytes.
+ */
+export async function* toLines(
+  buffs: AsyncIterable<Uint8Array>,
+): AsyncIterableIterator<string> {
+  const decoder = new TextDecoder();
+
+  let currentLine: Uint8Array[] = [];
+
+  const createLine = (): string => {
+    const line = concat(currentLine);
+    if (line.length > 0 && line[line.length - 1] === 0x0C) {
+      /* Strip the carriage return. */
+      return decoder.decode(line.slice(0, line.length - 1));
+    } else {
+      return decoder.decode(line);
+    }
+  };
+
+  for await (const buff of buffs) {
+    const length = buff.length;
+
+    let start = 0;
+    for (let pos = 0; pos < length; pos++) {
+      if (buff[pos] === 0x0A) {
+        if (pos) {
+          currentLine.push(buff.slice(start, pos));
+        }
+
+        yield createLine();
+
+        currentLine = [];
+        start = pos + 1;
+      }
+    }
+    if (start < length) {
+      currentLine.push(buff.slice(start));
+    }
+  }
+
+  if (currentLine.length > 0) {
+    yield createLine();
+  }
+}
+
+/**
  * Efficiently transform a `Reader` to an `AsyncIterableIterator<Uint8Array>`.
  * @param reader The reader.
  * @param bufSize The buffer size.
@@ -83,7 +130,32 @@ export async function* readerToBytes(
   }
 }
 
-export function concat(arrays: Uint8Array[]) {
+/**
+ * Transform a `Reader` to an `AsyncIterableIterator<Uint8Array>`, unbuffered.
+ * This will read bytes as soon as they are available.
+ *
+ * @param reader The reader.
+ * @param bufSize The buffer size.
+ */
+export async function* readerToBytesUnbuffered(
+  reader: Deno.Reader & Deno.Closer,
+): AsyncIterableIterator<Uint8Array> {
+  try {
+    const buffer = new Uint8Array(DEFAULT_BUFFER_SIZE);
+
+    while (true) {
+      const len = await reader.read(buffer);
+      if (len === null) {
+        break;
+      }
+      yield buffer.slice(0, len);
+    }
+  } finally {
+    reader.close();
+  }
+}
+
+export function concat(arrays: Uint8Array[]): Uint8Array {
   if (!arrays.length) return new Uint8Array(0);
 
   const totalLength = arrays.reduce((acc, value) => acc + value.length, 0);