+++ release-notes.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## 0.12.0
+
+- `#19` **enhancement** support for unbuffered input/output handlers (optional).
+- `#20` **enhancement** `stderr` is unbuffered.
+
 ## 0.11.1
 
 - `#17` **fix** `stringArrayInput()` performance.